requirements.txt
# flask
# transformers
# torch

Flask==2.3.3
Werkzeug==2.3.7
boto3==1.28.0
botocore==1.31.0
click==8.1.7
colorama==0.4.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
python-dotenv